commit to version 1.1

+++ provinder_8081/src/main/java/com/grace/it/service/impl/DeptManagerServiceImpl.java
@@ -43,7 +43,7 @@ public class DeptManagerServiceImpl implements DeptManagerService {
 
     @Transactional
     @Override
-    public Integer insert(DeptManager deptManager) {
+    public Long insert(DeptManager deptManager) {
         return deptManagerDao.insert(deptManager);
     }
